# Conflicts:
#	lib/Screens/challenge-detail.dart

+++ lib/Screens/task-detail.dart
@@ -43,9 +43,11 @@ class TaskDetail extends StatefulWidget {
 }
 
 class _TaskDetailState extends State<TaskDetail> {
+  // initialize needed variables
   List<ListTileModel> _items = [];
   final FirebaseAuth auth = FirebaseAuth.instance;
 
+  // get id of current logged in user
   String getUid() {
     final User? user = auth.currentUser;
     final uid = user!.uid;
@@ -53,17 +55,49 @@ class _TaskDetailState extends State<TaskDetail> {
     return uid.toString();
   }
 
+  Future<List> getAdminList() async {
+    String activeTeam = '';
+    List adminList = [];
+    String creator = '';
+    await FirebaseFirestore.instance
+        .collection('user')
+        .doc(getUid())
+        .get()
+        .then((value) => activeTeam = value['activeTeam']);
+    await FirebaseFirestore.instance
+        .collection('teams')
+        .doc(activeTeam)
+        .get()
+        .then((value) => creator = value['creator']);
+    await FirebaseFirestore.instance
+        .collection('teams')
+        .doc(activeTeam)
+        .get()
+        .then((value) => adminList = value['admins']);
+    adminList.add(creator);
+    return adminList;
+  }
+
+  Future<bool> checkAdmin() async {
+    List adminList = await getAdminList();
+      if (adminList.contains(getUid()))
+        return true;
+      else
+        return false;
+  }
+
+  // add subtasks to list
   void _add(String input) {
     _items.add(ListTileModel(false, input));
     setState(() {});
   }
-
   void initState() {
     for (int i = 0; i < widget.subTasks.length; i++) {
       _add(widget.subTasks[i]);
     }
   }
 
+  // get all subtasks
   Widget getSubtasks() {
     if (widget.subTasks.length != 0) {
       return new ListView(
@@ -115,15 +149,20 @@ class _TaskDetailState extends State<TaskDetail> {
                       child: Column(
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
+                          // task title
                           Text(widget.title,
                               style: TextStyle(
                                   fontWeight: FontWeight.bold, fontSize: 24)),
                           Padding(padding: EdgeInsets.all(10)),
+                          // task description
                           Text(widget.description,
                               style: TextStyle(fontSize: 18)),
+                          // Subtasks
                           getSubtasks(),
+                          // task xp
                           Text('XP: ' + widget.xp.toString()),
                           Padding(padding: EdgeInsets.all(10)),
+                          // upload time and date
                           Row(
                             children: [
                               Icon(Icons.access_time_outlined),
@@ -135,8 +174,11 @@ class _TaskDetailState extends State<TaskDetail> {
                           Padding(padding: EdgeInsets.all(10)),
                         ],
                       )),
+                  // Buttons depending on finished/accepted status
+                  // no button if finished by user
                   if (widget.finished && widget.userId == getUid())
                     SizedBox()
+                    // finish and cancel buttons if accepted by user
                   else if (widget.accepted && widget.userId == getUid())
                     Column(
                       mainAxisAlignment: MainAxisAlignment.center,
@@ -181,6 +223,7 @@ class _TaskDetailState extends State<TaskDetail> {
                                 child: Text('Abbrechen')))
                       ],
                     )
+                    // accept button if not finished or accepted
                   else if (!widget.accepted)
                     Container(
                         width: 300,
@@ -214,15 +257,29 @@ class _TaskDetailState extends State<TaskDetail> {
         ));
   }
 
-  void deleteTask() {
-    FirebaseFirestore.instance
-        .collection('tasks')
-        .doc(widget.id)
-        .delete();
-    Navigator.of(context).pop();
-    Navigator.of(context).pop();
+  // delete task from database
+  Future<void> deleteTask() async {
+    bool checkAdmins = await checkAdmin();
+    if (checkAdmins == true){
+      FirebaseFirestore.instance
+          .collection('tasks')
+          .doc(widget.id)
+          .delete();
+      Navigator.of(context).pop();
+      Navigator.of(context).pop();
+    } else {
+      print('Ist kein admin');
+      print(getUid());
+      final snackBar = SnackBar(
+          content: Text('Nur Admins können Tasks löschen.'));
+      ScaffoldMessenger.of(context)
+          .showSnackBar(snackBar);
+      Navigator.of(context).pop();
+      Navigator.of(context).pop();
+    }
   }
 
+  // set task as accepted in database
   void acceptTask(String id) {
     FirebaseFirestore.instance
         .collection('tasks')
@@ -231,6 +288,7 @@ class _TaskDetailState extends State<TaskDetail> {
     Navigator.of(context).pop();
   }
 
+  // set task as finished in database and check if user has levelup
   Future<void> finishTask(String id) async {
     var task = await FirebaseFirestore.instance
         .collection('tasks')
@@ -246,9 +304,11 @@ class _TaskDetailState extends State<TaskDetail> {
         .collection('user')
         .doc(getUid());
 
+    // check for levelup
     final CalculateLevel logic = new CalculateLevel();
     logic.levelUp(user);
 
+    // update user in database
     user.update({'finishedTasksCount': FieldValue.increment(1)});
     user.update({'xp': FieldValue.increment(taskXp)});
 
@@ -256,6 +316,7 @@ class _TaskDetailState extends State<TaskDetail> {
     Navigator.pushReplacementNamed(context, 'tasks');
   }
 
+  // set task as not accepted in database
   void cancelTask(String id) {
     FirebaseFirestore.instance
         .collection('tasks')
